# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/articles/": {
            "get": {
                "description": "获取文章列表",
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "创建文章",
                "parameters": [
                    {
                        "description": "body for Culture content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/articles.ArticleModelValidator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/user": {
            "get": {
                "description": "获取用户信息",
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户信息",
                "parameters": [
                    {
                        "description": "body for Culture content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/users.UserModelValidator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/": {
            "post": {
                "description": "注册用户",
                "parameters": [
                    {
                        "description": "body for Culture content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/users.UserModelValidator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "description": "用户登录",
                "parameters": [
                    {
                        "description": "body for Culture content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/users.LoginValidator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{username}/follow": {
            "delete": {
                "description": "取消关注用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\": []*TypeResponse,\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "articles.ArticleModelValidator": {
            "type": "object",
            "properties": {
                "article": {
                    "type": "object",
                    "properties": {
                        "body": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "tagList": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "title": {
                            "type": "string"
                        },
                        "typeId": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "users.LoginValidator": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "email": {
                            "type": "string"
                        },
                        "password": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "users.UserModelValidator": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "bio": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "image": {
                            "type": "string"
                        },
                        "password": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}